Data Wrangling, Training data, Best K
Wrangled the data so that the Age variable is now named age. Additionally, used drop_na() to remove all the NAs in the age variable (this fixed the issue that I was experiencing when trying to find best k). Set the seed and split the dataset to train and test. Created recipe to standardize the training dataset. Performed cross validation and chose the K that gave best accuracy. Then finally began training the classifier. Undetailed explanations written as a place holder for now.